# misc performance improvements