--- a/src/vec.h
+++ b/src/vec.h
@@ -26,6 +26,10 @@
 
 namespace manifold {
 
+#if (MANIFOLD_PAR == 1)
+extern tbb::task_arena gc_arena;
+#endif
+
 template <typename T>
 class Vec;
 
@@ -92,8 +96,7 @@ class Vec : public VecView<T> {
 
   ~Vec() {
     if (this->ptr_ != nullptr) {
-      TracyFreeS(this->ptr_, 3);
-      free(this->ptr_);
+      free_async(this->ptr_, capacity_);
     }
     this->ptr_ = nullptr;
     this->size_ = 0;
@@ -103,8 +106,7 @@ class Vec : public VecView<T> {
   Vec<T>& operator=(const Vec<T>& other) {
     if (&other == this) return *this;
     if (this->ptr_ != nullptr) {
-      TracyFreeS(this->ptr_, 3);
-      free(this->ptr_);
+      free_async(this->ptr_, capacity_);
     }
     this->size_ = other.size_;
     capacity_ = other.size_;
@@ -120,8 +122,7 @@ class Vec : public VecView<T> {
   Vec<T>& operator=(Vec<T>&& other) {
     if (&other == this) return *this;
     if (this->ptr_ != nullptr) {
-      TracyFreeS(this->ptr_, 3);
-      free(this->ptr_);
+      free_async(this->ptr_, capacity_);
     }
     this->size_ = other.size_;
     capacity_ = other.capacity_;
@@ -166,8 +167,7 @@ class Vec : public VecView<T> {
         manifold::copy(autoPolicy(this->size_), this->ptr_,
                        this->ptr_ + this->size_, newBuffer);
       if (this->ptr_ != nullptr) {
-        TracyFreeS(this->ptr_, 3);
-        free(this->ptr_);
+        free_async(this->ptr_, capacity_);
       }
       this->ptr_ = newBuffer;
       capacity_ = n;
@@ -208,8 +208,7 @@ class Vec : public VecView<T> {
       manifold::copy(this->ptr_, this->ptr_ + this->size_, newBuffer);
     }
     if (this->ptr_ != nullptr) {
-      TracyFreeS(this->ptr_, 3);
-      free(this->ptr_);
+      free_async(this->ptr_, capacity_);
     }
     this->ptr_ = newBuffer;
     capacity_ = this->size_;
@@ -221,5 +220,19 @@ class Vec : public VecView<T> {
   size_t capacity_ = 0;
 
   static_assert(std::is_trivially_destructible<T>::value);
+
+  static void free_async(T* ptr, size_t size) {
+    TracyFreeS(ptr, 3);
+    // only do async free if the size is large, because otherwise we may be able
+    // to reuse the allocation, and the deallocation probably won't trigger
+    // munmap.
+    // Currently it is set to 64 pages.
+#if (MANIFOLD_PAR == 1)
+    if (size * sizeof(T) > 1 << 18)
+      gc_arena.enqueue([ptr]() { free(ptr); });
+    else
+#endif
+      free(ptr);
+  }
 };
 }  // namespace manifold