Misc changes, they are mainly optimized for large meshes and have small to no impact to smaller meshes which are the majority of our benchmarks, so I am skipping them here.

Old perf:

```
nTri = 512, time = 0.00102039 sec
nTri = 2048, time = 0.00176623 sec
nTri = 8192, time = 0.00433313 sec
nTri = 32768, time = 0.00954862 sec
nTri = 131072, time = 0.0269745 sec
nTri = 524288, time = 0.107106 sec
nTri = 2097152, time = 0.492739 sec
nTri = 8388608, time = 1.93114 sec
```



1. Allow the collider to take a function for queries. This avoids the need to construct a `TmpEdge` and `Box` vector for `Intersect12`, which makes things slightly faster when the mesh is particularly huge.
  
  ```
  nTri = 512, time = 0.00129358 sec
  nTri = 2048, time = 0.0016819 sec
  nTri = 8192, time = 0.0040756 sec
  nTri = 32768, time = 0.00914949 sec
  nTri = 131072, time = 0.0252436 sec
  nTri = 524288, time = 0.0948401 sec
  nTri = 2097152, time = 0.407921 sec
  nTri = 8388608, time = 1.66111 sec
  ```

2. Free large vectors asynchronously. Large allocations are typically treated differently by the allocator, they are requested from and released to the OS directly. For allocation, there isn't much we can do unless we reuse the buffer (this will be fun to try). However, for deallocation, we can just move it to a separate thread to do all the munmap work, improving performance. This one is quite significant:

  ```
  nTri = 512, time = 0.0010744 sec
  nTri = 2048, time = 0.00183495 sec
  nTri = 8192, time = 0.00467922 sec
  nTri = 32768, time = 0.00894273 sec
  nTri = 131072, time = 0.0235959 sec
  nTri = 524288, time = 0.0874061 sec
  nTri = 2097152, time = 0.376126 sec
  nTri = 8388608, time = 1.51217 sec
  ```